% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_postgres.R
\name{connection_postgres}
\alias{connection_postgres}
\title{connection_postgres}
\usage{
connection_postgres(
  POSTGRES_HOST = NULL,
  POSTGRES_PORT = NULL,
  POSTGRES_USER = NULL,
  POSTGRES_PASSWORD = NULL,
  POSTGRES_DB = NULL,
  n_connection_tries = 5
)
}
\arguments{
\item{POSTGRES_HOST}{Database hostname}

\item{POSTGRES_PORT}{Database port}

\item{POSTGRES_PASSWORD}{Database password}

\item{POSTGRES_DB}{Database name}

\item{n_connection_tries}{Number of attempts to connect}
}
\description{
Adds the ability to re-try connections on failure and reads from pre-defined environment variables names.
}
