% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_variable_check.R
\name{env_variable_check}
\alias{env_variable_check}
\title{env_variable_check}
\usage{
env_variable_check(env_variable)
}
\description{
env_variable check verifies that a variable
coming from the .Renviron file actually exists.
}
