% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello
#' @author James King}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{Someone's name}
}
\description{
hello
#' @author James King
}
\examples{
hello(name = 'James')
}
