% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_postgres.R
\name{connection_postgres}
\alias{connection_postgres}
\title{connection_postgres}
\usage{
connection_postgres(
  POSTGRES_HOST = NULL,
  POSTGRES_PORT = NULL,
  POSTGRES_USER = NULL,
  POSTGRES_PASSWORD = NULL,
  POSTGRES_DB = NULL
)
}
\description{
connection_postgres
}
