% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_postgres.R
\name{connection_postgres}
\alias{connection_postgres}
\title{connection_postgres}
\usage{
connection_postgres(
  POSTGRES_HOST = NULL,
  POSTGRES_PORT = NULL,
  POSTGRES_USER = NULL,
  POSTGRES_PASSWORD = NULL,
  POSTGRES_DB = NULL,
  n_connection_tries = 5
)
}
\arguments{
\item{POSTGRES_HOST}{for the host name}

\item{POSTGRES_PORT}{for the port number}

\item{POSTGRES_USER}{for the user name}

\item{POSTGRES_PASSWORD}{for the password}

\item{POSTGRES_DB}{for the name of the database on the host, or the database file name}

\item{n_connection_tries}{Number of times to try connecting to PostgreSQL}
}
\description{
Adds the ability to re-try connections on failure and reads from
pre-defined environment variables names.
}
