% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{Enter a name that you want to print}
}
\description{
hello
}
\examples{
hello(name = "Richard")
}
\author{
Richard Mills
}
